Extract common code to "utils.hpp"
This will allow us to use that code in both "juce6.hpp",
"reprojucer.hpp", and "main.cpp" in subsequent commits.
